GrindrPlus: Add functionality to replace the Maps API Key
This new feature allows the user to specify a custom API Key for Google Maps when using Grindr Plus with LSPatch. The original API key used by Grindr is restricted to the original package signature. When using LSPatch, the signature changes, which impairs the functionality of features such as Explore, location sharing and the teleport command. API keys can be created using the Google Cloud Console. Providing a credit card is required, however, the free quota should be more than enough for a single user. Using an unrestricted API key is probably the easiest solution as there is currently no logic to display the new signature and the signing key is also very short-lived (30 days).