GrindrPlus: Rename `ExpiringPhotos` to `ExpiringMedia`
This refactors the existing Unlimited Expiring Pictures hook with a revamped
implementation that saves images directly to the filesystem instead of using
a database. This commit also adds initial support for expiring videos (there
is still some work to be done before this is complete).

The new methodology stores media in the app's files directory using the media
IDs. From previous experiments, the files are always converted to either JPG
(for images) or MP4 (for videos), so these are the default formats, though
support for other formats has been included for future compatibility.

Change-Id: I0a3d6f009ee27a32a060bb35f5795c02e40c2296