GrindrPlus: Update the geohash when teleporting
Initially, this was handled by the Grindr app after teleporting and
fetching the cascade. However, this approach often led to "Unable to
determine your location" errors due to the slow response time of the
GPS API on Android.

To resolve this, we should instead convert the desired coordinates
directly into a geohash and send that to the Grindr API. This way,
the app only needs to refresh the cascade, avoiding GPS-related delays.

Change-Id: I82b5809b255ff5da169692154909cfb8d78e973f