GrindrPlus: Override app restart dialog
The goal of this change is to help me identify why certain users have
specific features blocked while others don't. This behavior is likely
influenced by location or a similar factor. In any case, displaying
the stack trace will help me pinpoint the root cause of the issue.

Additionally, this will reduce confusion, as the original dialog was
telling users that an app restart was required, which is not the case.

Change-Id: I97ab152a3d1c13ce440f82b04ec1026a2585d6ff