GrindrPlus: Make bridge service a foreground service
This helps keep the bridge service running even when the Grindr app is pushed
to the background by the system, preventing unwanted disconnections.

The only downside is that it shows a persistent notification to let users know
a foreground service is active, but that can be turned off in settings if needed.

It also makes the app launch a bit quicker, since the connection between both the
bridge client and the service is already in place.

Change-Id: Ie75443340b1df8785c50d02ea0cff9c79b9c4406