GrindrPlus: Introduce background tasks
This commit adds a new background task manager that lets GrindrPlus schedule,
configure, and run multiple background tasks, even when the app isn't open on
the main thread. This system makes it possible to support features that don't
rely on the app being in the foreground.

A basic proof-of-concept task is included, which keeps your account persistently
online so it continues to appear on the grid. Since this behavior can be suspicious
and may lead to account bans, all tasks are turned off by default. Users can choose
to enable them manually if they understand the risks.

Change-Id: I2923cedd8d7d8eecfe1302164ceb10eba387f8b2